Cleaning up markup and typescript
