+++ src/components/navigation-header/components/index.ts
@@ -1,0 +1,3 @@
+import NavigationHeaderDropdownMenu from './dropdown-menu/dropdown-menu'
+
+export { NavigationHeaderDropdownMenu }
